payments, verification and notifications

+++ src/main/java/org/noctisdev/sciallhexvsg/payments/domain/external/IPayments.java
@@ -1,5 +1,5 @@
 package org.noctisdev.sciallhexvsg.payments.domain.external;
 
 public interface IPayments {
-    String generateDonationPaymentUrl(String to, Float amount);
+    String generateDonationPaymentUrl(String to, Double amount);
 }
